Support typeFullName for cast/index access via rust_ast_gen.
